# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/reviews/product/{productId}": {
      "get": {
        "operationId": "ReviewsController_findByProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "query",
            "description": "ID del producto",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/reviews": {
      "post": {
        "operationId": "ReviewsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/reviews/{id}": {
      "delete": {
        "operationId": "ReviewsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/reviews/admin/all": {
      "get": {
        "operationId": "ReviewsController_findAllAdmin",
        "parameters": [
          {
            "name": "includeInactive",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/reviews/{id}/restore": {
      "patch": {
        "operationId": "ReviewsController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReviewUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "4539152d-031b-41c4-8ef8-c419e8b5f84a"
          },
          "name": {
            "type": "string",
            "example": "Juan Pérez",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "product_id": {
            "type": "number",
            "example": 1
          },
          "user_id": {
            "type": "string",
            "example": "4539152d-031b-41c4-8ef8-c419e8b5f84a"
          },
          "rating": {
            "type": "number",
            "example": 5,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "example": "Excelente producto",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "example": "2026-04-10T00:00:00.000Z",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "users": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewUser"
              }
            ]
          }
        },
        "required": [
          "id",
          "product_id",
          "user_id",
          "rating",
          "comment",
          "created_at",
          "is_active",
          "users"
        ]
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "number",
            "x-nestjs_zod-has-null": true
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "x-nestjs_zod-has-null": true
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-has-null": true,
            "x-nestjs_zod-empty-type": true
          }
        },
        "required": [
          "product_id",
          "rating"
        ]
      },
...
    }
  }
}